refactor user access helper
