+++ src/common/interfaces/auth-user.interface.ts
@@ -25,19 +25,3 @@ export interface UserAccess {
   role: string[];
   isAdmin: boolean;
 }
-
-export const getUserAccess = (authUser: AuthUser): UserAccess => {
-  const role = authUser.resource_access['digiteam-inventaris-app'].roles.map(
-    (role) => role,
-  );
-
-  const isAdmin =
-    authUser.resource_access['digiteam-inventaris-app'].roles.includes('admin');
-
-  return {
-    name: authUser.name,
-    email: authUser.email,
-    role: role,
-    isAdmin: isAdmin,
-  };
-};
